pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "fastapi>=0.125.0",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "python-multipart>=0.0.21",
    "sqlalchemy[asyncio]>=2.0.45",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
